+++ src/ball/ball.c
@@ -6,12 +6,12 @@
 #include "../../include/display.h"
 
 void init_ball(Ball* ball) {
-    ball->dx = 4;
-    ball->dy = 4;
-    ball->x = BUFFER_W / 2;
+    ball->dx = 4.0;
+    ball->dy = 4.0;
+    ball->x = BUFFER_W / 2.0;
     ball->y = BUFFER_H - PAD_H - BALL_R;
 }
 
-void draw_ball(int ball_x, int ball_y) {
+void draw_ball(float ball_x, float ball_y) {
     al_draw_filled_circle(ball_x, ball_y, BALL_R, al_map_rgb_f(1, 1, 0));
 }
